move getPdf to scripts dir
