+++ .gitignore
@@ -13,7 +13,6 @@
 /out/
 
 # production
-/build
 /public/adam-drago-resume.pdf
 
 # misc
