+++ scripts/getPdf.ts
